requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
asyncpg==0.29.0

# Message Queue
redis==5.0.1

# DNS and Domain Tools
dnspython==2.4.2
python-whois==0.8.0

# GitHub Integration
PyGithub==2.1.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and Monitoring
loguru==0.7.2
